Build fix for newer g++: enum != int
git-svn-id: http://astade.tigris.org/svn/astade/trunk@1917 aaa30fad-34f9-0310-8aa9-877cc762ee2e

+++ Model/package_SVSWQPEV/Package_EWMKSFJL/Package_HMPGOZXK/classes/class_JLNJGYOF/operations/operation_ARGCABNQ/code.cpp
@@ -1,4 +1,4 @@
 configObject.Write("Label",myLabel.myText);
-configObject.Write("Type",type);
+configObject.Write("Type",static_cast<long>(type));
 configObject.Write("StartMult",myStartMult.myText);
 configObject.Write("EndMult",myEndMult.myText);
